requirements.txt
asgiref==3.7.2
boto3==1.28.57
botocore==1.31.57
cffi==1.16.0
cryptography==41.0.4
Django==4.2.5
django-cache==0.3
django-cors-headers==4.2.0
django-environ==0.11.2
django-redis==5.3.0
django-storages==1.14.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jmespath==1.0.1
mysqlclient==2.2.0
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
